requirements.txt
fastapi
uvicorn
python-multipart
requests
PyMuPDF
python-docx
pytesseract
Pillow
pandas
langdetect
whoosh